Mobile and componentization of Donate and Blog Cards
